lwm2m: fix bug commit validate keys
